requirements.txt
Pillow==10.2.0
uvicorn==0.29.0
python-multipart==0.0.9
fastapi==0.110.1
nest-asyncio==1.6.0
pandas==2.2.2
numpy==1.25
ipykernel
requests==2.32.3
python_dotenv==1.0.1
plotly
seaborn==0.13.2
matplotlib==3.9.2
beautifulsoup4==4.12.2
selenium==4.10.0
sentencepiece
sentence_transformers==3.2.0
sacremoses==0.1.1
pyairtable~=3.2.0
openpyxl==3.1.5
google_api_python_client==2.185.0
google-auth
google-auth-oauthlib~=1.2.2
xlsxwriter==3.2.9
python-dotenv~=1.0.1
google-api-python-client~=2.185.0
protobuf~=6.33.0
transformers~=4.57.1